requirements.txt
annotated-types==0.6.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
deprecation==2.1.0
filelock==3.13.3
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.6
lancedb==0.6.7
numpy==1.26.4
overrides==7.7.0
packaging==24.0
pillow==10.3.0
py==1.11.0
pyarrow==13.0.0
pydantic==2.6.4
pydantic_core==2.16.3
pylance==0.10.9
PyYAML==6.0.1
ratelimiter==1.2.0.post0
requests==2.31.0
retry==0.9.2
semver==3.0.2
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.11.0
uform==2.0.2
urllib3==2.2.1
